requirements.txt
fastapi
pydantic
uvicorn
SQLAlchemy
python-dotenv
asyncpg
aiomysql
geoalchemy2
requests
redis
celery